Create rpi_read_temp.c
This file measures temperature from a sensor. It 
configures the Serial Peripheral Unit (SPI) in 
the micro-controller to communicate with a 
8 channel Analog to Digital Converter 
(ADC - MCP3008). The ADC reads
the voltage drop across a thermistor and sends 
the voltage as counts. The software scales the
counts to voltage based on the ADC resolution.
The software then uses a look up table to convert
the voltage to temperature.